refactor: ScheduleTemplate 생성 로직을 별도 패키지로 분리
