+++ src/main/java/com/burntoburn/easyshift/service/templates/ShiftTemplateService.java
@@ -1,4 +1,4 @@
-package com.burntoburn.easyshift.service.shift;
+package com.burntoburn.easyshift.service.templates;
 
 import com.burntoburn.easyshift.entity.templates.ShiftTemplate;
 import java.util.List;
